+++ main.go
@@ -57,7 +57,7 @@ func (rr *replReader) ReadLine() (string, bool) {
 // Version is bumped at release time. Override at build with:
 //
 //	go build -ldflags "-X main.Version=v0.2.0"
-var Version = "v0.74.0"
+var Version = "v0.75.0"
 
 const (
 	cReset  = "\033[0m"